Move remove_dead_funcs stuff into separate file
